+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Portable Abstractions and Patterns for Golang Changelog
 
+## <a name="1.1.4"></a> 1.1.4 (2021-12-06)
+### Bug Fixing 
+* Fixed TypeConverter, ToTypeWithDefault method with nil default value crashed app
+* Fixed AnyValueArray, GetAsMapWithDefault crashed app with stack overflow error
+
 ## <a name="1.1.3"></a> 1.1.3 (2021-10-22)
 ### Features 
 * Added support access to object field by names in tags in reflect package